Add React Stepper to Front-End
Added the @mui/material, @emotion/react, @emotion/styled
packages to the React project so that I would be able to
use the stepper.

Implemented the stepper in the Quiz.tsx. Made it so that
each question is in a form and the key and value is extracted
to the formData variable which is later sent to the backEnd
for processing.

Exported the Question interface to a seperate
file in the interface folder.

Updated the quiz post so that it recieves
json data.