Fix double sending of data in Quiz front-end
Before of setFormData an additional post request would
be triggered. Fixed that by removing the setFormData
from around the post request and adding updatedFormData
to get the latest data to submit (if not done, the last questions
answer may be lost).

Extracted the logic from handleNext to getUpdateData so
it may be used in handleSubmit.

Made the submit button green.